Create 1_Pandas_Project_A\B_Testing_for_ShoeFly.com
Our favorite online shoe store, ShoeFly.com is performing an A/B Test. They have two different versions of an ad, which they have placed in emails, as well as in banner ads on Facebook, Twitter, and Google. They want to know how the two ads are performing on each of the different platforms on each day of the week. Help them analyze the data using aggregate measures.
Your manager wants to know which ad platform is getting you the most views.
Create a new column called is_click, which is True if ad_click_timestamp is not null and False otherwise.
We want to know the percent of people who clicked on ads from each utm_source and pivot the data.
Create a new column in clicks_pivot called percent_clicked which is equal to the percent of users who clicked on the ad from each utm_source.
Were approximately the same number of people shown both adds A and B?
Check to see if a greater percentage of users clicked on Ad A or Ad B.
The Product Manager for the A/B test thinks that the clicks might have changed by day of the week.
Do you recommend that your company use Ad A or Ad B?

Przeanalizować dane ze sklepu ShoeFly.com związane z reklamami i użytkownikami, którzy kliknęli w poszczególne źródła reklamowe.