Create 2_Pandas_Project_Page Visits Funnel.py
Cool T-Shirts Inc. has asked you to analyze data on visits to their website. Your job is to build a funnel, which is a description of how many people continue to the next step of a multi-step process.
In this case, our funnel is going to describe the following process:
1. A user visits CoolTShirts.com
2. A user adds a t-shirt to their cart
3. A user clicks “checkout”
4. A user actually purchases a t-shirt

How long is your merged DataFrame?
How many of the timestamps are null for the column cart_time?
What percent of users who visited Cool T-Shirts Inc. ended up not placing a t-shirt in their cart?
What percentage of users put items in their cart, but did not proceed to checkout?
Merge all four steps of the funnel, in order, using a series of left merges.
What percentage of users proceeded to checkout, but did not purchase a t-shirt?
Which step of the funnel is weakest (i.e., has the highest percentage of users not completing it)?
Using the giant merged DataFrame all_data that you created, let’s calculate the average time from initial visit to final purchase.